Port tendril benchmarks to criterion
